requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
groq>=0.31.0
python-dotenv>=1.0.0